// mobile/src/models/sub_summaries.model.ts
import z from "zod";

export const SubSummariesSchema = z.object({
	id: z.string(),
	summary_id: z.string(),
	name: z.string(),
	total: z.string(),
});

// insertion schema (id generated by SQLite / ULID)
export const CreateSubSummariesSchema = SubSummariesSchema.omit({
	id: true,
});

export type SubSummary = z.infer<typeof SubSummariesSchema>;

export type CreateSubSummary = z.infer<typeof CreateSubSummariesSchema>;